Add `MigrationRuntimeSupport` to improve runtime tracking and audit operations
- Introduced `MigrationRuntimeDetails` for enhanced migration status tracking, including phase, attempt, ETA, and error details.
- Added utility functions for projecting migration runtime status, updating runtime details, and managing audit events and metrics.
- Refactored FoundationDB migration handling to utilize new runtime details, metrics, and audit logging.
